dpdk: enable RSS hash calculation also on 1 NIC queue
